+++ src/components/burger-constuctor/burger-constructor.module.css
@@ -14,6 +14,7 @@
   gap: 16px;
   overflow-y: scroll;
   scroll-behavior: smooth;
+  width: 605px; 
 }
 
 .bread {
