+++ src/components/ingredient-details/ingredient-details.jsx
@@ -1,5 +1,4 @@
 import styles from './ingredient-details.module.css';
-//import { ingredientPropType } from "../../utils/prop-types";
 import { useSelector } from 'react-redux';
 
 const IngredientDetails = () =>{
