--- src/services/actions/burger-ingredients.js
@@ -1,8 +1,0 @@
-export const ITEM_OPEN = 'ITEM_OPEN';
-export const ITEM_CLOSE = 'ITEM_CLOSE';
-export const MODAL_OPEN = 'MODAL_OPEN';
-export const MODAL_CLOSE = 'MODAL_CLOSE';
-export const BUN_MOVE ='BUN_MOVE';
-export const SAUCE_FILLING_MOVE ='SAUCE_FILLING_MOVE';
-export const ELEMENT_REMOVE ='ELEMENT_REMOVE';
-export const SORT_INGREDIENTS = 'SORT_INGREDIENTS';
