--- src/services/actions/index.js
@@ -1,1 +1,0 @@
-// здесь будет код
